third_party/WebKit/LayoutTests/svg/custom/svg-parse-overflow-4-expected.txt

CONSOLE ERROR: line 9: Error: <foreignObject> attribute y: Expected length, "105.0e50".
